requirements.txt
python-dotenv
requests
pandas
numpy
tqdm
langdetect
scikit-learn
matplotlib
seaborn
nltk
jupyter
